[Templating] ease migration with template helpers (#7463)
* [Templating] ease migration with template helpers by adding them to 6.8 already and alias them to php template helpers. this also makes it possible to make bundles compatible with 6.8 and 7.0

* [Templating] ease migration with template helpers by adding them to 6.8 already and alias them to php template helpers. this also makes it possible to make bundles compatible with 6.8 and 7.0 #2 fix CacheBuster Stan issue

* [Templating] ease migration with template helpers by adding them to 6.8 already and alias them to php template helpers. this also makes it possible to make bundles compatible with 6.8 and 7.0 #baseline

* [Templating] ease migration with template helpers by adding them to 6.8 already and alias them to php template helpers. this also makes it possible to make bundles compatible with 6.8 and 7.0 #baseline

* Update lib/Templating/Helper/Navigation/Exception/RendererNotFoundException.php
