+++ bundles/CoreBundle/Resources/config/templating_twig.yml
@@ -37,6 +37,45 @@ services:
 
     Pimcore\Twig\Extension\DumpExtension: ~
 
+    Pimcore\Twig\Extension\Templating\Action:
+        calls:
+            - [ setRoutingDefaults, [ '%pimcore.routing.defaults%' ] ]
+
+    Pimcore\Twig\Extension\Templating\PimcoreUrl:
+        class: Pimcore\Twig\Extension\Templating\PimcoreUrl
+        arguments:
+            $generator: '@router'
+
+    Pimcore\Twig\Extension\Templating\HeadLink: ~
+
+    Pimcore\Twig\Extension\Templating\HeadMeta: ~
+
+    Pimcore\Twig\Extension\Templating\HeadScript: ~
+
+    Pimcore\Twig\Extension\Templating\HeadStyle: ~
+
+    Pimcore\Twig\Extension\Templating\HeadTitle: ~
+
+    Pimcore\Twig\Extension\Templating\Inc: ~
+
+    Pimcore\Twig\Extension\Templating\InlineScript: ~
+
+    Pimcore\Twig\Extension\Templating\Navigation:
+        lazy: true
+        arguments:
+            $rendererLocator: '@pimcore.templating.navigation.renderer_locator'
+
+    Pimcore\Twig\Extension\Templating\Placeholder: ~
+
+    Pimcore\Twig\Extension\Templating\Placeholder\AbstractExtension: ~
+
+    Pimcore\Twig\Extension\Templating\Placeholder\ContainerService: ~
+
+    Pimcore\Twig\Extension\Templating\Placeholder\Container: ~
+
+    Pimcore\Twig\Extension\Templating\Placeholder\CacheBusterAware: ~
+
+
     # the deferred extension is needed for placeholder helpers to work
     # as otherwise the placeholder block would be rendered before any
     # content was added (e.g. headTitle)
@@ -45,3 +84,4 @@ services:
     # provides truncate and wordwrap filters
     Twig_Extensions_Extension_Text:
         class: Twig_Extensions_Extension_Text
+
