+++ lib/Templating/Helper/Action.php
@@ -1,4 +1,7 @@
 <?php
+
+declare(strict_types=1);
+
 /**
  * Pimcore
  *
@@ -14,84 +17,19 @@
 
 namespace Pimcore\Templating\Helper;
 
-use Pimcore\Model\Document\PageSnippet;
-use Pimcore\Targeting\Document\DocumentTargetingConfigurator;
-use Pimcore\Templating\Renderer\ActionRenderer;
-use Symfony\Component\Templating\Helper\Helper;
+@trigger_error(
+    'Pimcore\Templating\Helper\Action is deprecated since version 6.8.0 and will be removed in 7.0.0. ' .
+    ' Use ' . \Pimcore\Twig\Extension\Templating\Action::class . ' instead.',
+    E_USER_DEPRECATED
+);
 
-/**
- * @deprecated
- */
-class Action extends Helper
-{
+class_exists(\Pimcore\Twig\Extension\Templating\Action::class);
+
+if (false) {
     /**
-     * @var ActionRenderer
+     * @deprecated since Pimcore 6.8, use Pimcore\Twig\Extension\Templating\Action
      */
-    protected $actionRenderer;
+    class Action extends \Pimcore\Twig\Extension\Templating\Action {
 
-    /**
-     * @var DocumentTargetingConfigurator
-     */
-    private $targetingConfigurator;
-
-    /**
-     * @var array
-     */
-    private $routingDefaults = [];
-
-    public function __construct(
-        ActionRenderer $actionRenderer,
-        DocumentTargetingConfigurator $targetingConfigurator
-    ) {
-        $this->actionRenderer = $actionRenderer;
-        $this->targetingConfigurator = $targetingConfigurator;
-    }
-
-    public function setRoutingDefaults(array $routingDefaults)
-    {
-        $this->routingDefaults = $routingDefaults;
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function getName()
-    {
-        return 'action';
-    }
-
-    /**
-     * @param string $action
-     * @param string $controller
-     * @param string|null $module
-     * @param array $attributes
-     * @param array $query
-     * @param array $options
-     *
-     * @return string
-     */
-    public function __invoke($action, $controller, $module = null, array $attributes = [], array $query = [], array $options = [])
-    {
-        $document = isset($attributes['document']) ? $attributes['document'] : null;
-        if ($document && $document instanceof PageSnippet) {
-            // apply best matching target group (if any)
-            $this->targetingConfigurator->configureTargetGroup($document);
-
-            $attributes = $this->actionRenderer->addDocumentAttributes($document, $attributes);
-        }
-
-        if (!$module) {
-            $module = $this->routingDefaults['bundle'];
-        }
-
-        $uri = $this->actionRenderer->createControllerReference(
-            $module,
-            $controller,
-            $action,
-            $attributes,
-            $query
-        );
-
-        return $this->actionRenderer->render($uri, $options);
     }
 }
