+++ lib/Templating/Helper/Inc.php
@@ -1,4 +1,7 @@
 <?php
+
+declare(strict_types=1);
+
 /**
  * Pimcore
  *
@@ -14,58 +17,19 @@
 
 namespace Pimcore\Templating\Helper;
 
-use Pimcore\Http\Request\Resolver\EditmodeResolver;
-use Pimcore\Model\Document\PageSnippet;
-use Pimcore\Templating\Renderer\IncludeRenderer;
-use Symfony\Component\Templating\Helper\Helper;
+@trigger_error(
+    'Pimcore\Templating\Helper\Inc is deprecated since version 6.8.0 and will be removed in 7.0.0. ' .
+    ' Use ' . \Pimcore\Twig\Extension\Templating\Inc::class . ' instead.',
+    E_USER_DEPRECATED
+);
 
-/**
- * @deprecated
- */
-class Inc extends Helper
-{
+class_exists(\Pimcore\Twig\Extension\Templating\Inc::class);
+
+if (false) {
     /**
-     * @var IncludeRenderer
+     * @deprecated since Pimcore 6.8, use Pimcore\Twig\Extension\Templating\Inc
      */
-    protected $includeRenderer;
+    class Inc extends \Pimcore\Twig\Extension\Templating\Inc {
 
-    /**
-     * @var EditmodeResolver
-     */
-    protected $editmodeResolver;
-
-    /**
-     * @param IncludeRenderer $includeRenderer
-     * @param EditmodeResolver $editmodeResolver
-     */
-    public function __construct(IncludeRenderer $includeRenderer, EditmodeResolver $editmodeResolver)
-    {
-        $this->includeRenderer = $includeRenderer;
-        $this->editmodeResolver = $editmodeResolver;
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function getName()
-    {
-        return 'inc';
-    }
-
-    /**
-     * @param PageSnippet|int|string $include
-     * @param array $params
-     * @param bool $cacheEnabled
-     * @param bool|null $editmode
-     *
-     * @return string
-     */
-    public function __invoke($include, array $params = [], $cacheEnabled = true, $editmode = null)
-    {
-        if (null === $editmode) {
-            $editmode = $this->editmodeResolver->isEditmode();
-        }
-
-        return $this->includeRenderer->render($include, $params, $editmode, $cacheEnabled);
     }
 }
