+++ lib/Templating/Helper/Navigation/Exception/InvalidRendererException.php
@@ -19,23 +19,21 @@ namespace Pimcore\Templating\Helper\Navigation\Exception;
 
 use Pimcore\Navigation\Renderer\RendererInterface;
 
-class InvalidRendererException extends \LogicException
-{
-    /**
-     * @param string $name
-     * @param mixed $renderer
-     *
-     * @return InvalidRendererException
-     */
-    public static function create(string $name, $renderer): self
-    {
-        $type = is_object($renderer) ? get_class($renderer) : gettype($renderer);
+@trigger_error(
+    'Pimcore\Templating\Helper\Navigation\Exception\InvalidRendererException is deprecated since version 6.8.0 and will be removed in 7.0.0. ' .
+    ' Use ' . \Pimcore\Twig\Extension\Templating\Navigation\Exception\InvalidRendererException::class . ' instead.',
+    E_USER_DEPRECATED
+);
 
-        return new static(sprintf(
-            'Renderer for name "%s" was expected to implement interface "%s", "%s" given.',
-            $name,
-            RendererInterface::class,
-            $type
-        ));
+class_exists(\Pimcore\Twig\Extension\Templating\Navigation\Exception\InvalidRendererException::class);
+
+if (false) {
+    /**
+     * @deprecated since Pimcore 6.8, use Pimcore\Twig\Extension\Templating\Navigation\Exception\InvalidRendererException
+     */
+    class InvalidRendererException extends \Pimcore\Twig\Extension\Templating\Navigation\Exception\InvalidRendererException {
+
     }
 }
+
+
