+++ lib/Templating/Helper/Placeholder.php
@@ -1,4 +1,7 @@
 <?php
+
+declare(strict_types=1);
+
 /**
  * Pimcore
  *
@@ -12,75 +15,21 @@
  * @license    http://www.pimcore.org/license     GPLv3 and PEL
  */
 
-/**
- * ----------------------------------------------------------------------------------
- * based on @author ZF1 Zend_View_Helper_Placeholder
- * ----------------------------------------------------------------------------------
- */
-
-/**
- * Zend Framework
- *
- * LICENSE
- *
- * This source file is subject to the new BSD license that is bundled
- * with this package in the file LICENSE.txt.
- * It is also available through the world-wide-web at this URL:
- * http://framework.zend.com/license/new-bsd
- * If you did not receive a copy of the license and are unable to
- * obtain it through the world-wide-web, please send an email
- * to license@zend.com so we can send you a copy immediately.
- *
- * @copyright  Copyright (c) 2005-2015 Zend Technologies USA Inc. (http://www.zend.com)
- * @license    http://framework.zend.com/license/new-bsd     New BSD License
- */
-
 namespace Pimcore\Templating\Helper;
 
-use Pimcore\Templating\Helper\Placeholder\AbstractHelper;
-use Pimcore\Templating\Helper\Placeholder\Container;
+@trigger_error(
+    'Pimcore\Templating\Helper\Placeholder is deprecated since version 6.8.0 and will be removed in 7.0.0. ' .
+    ' Use ' . \Pimcore\Twig\Extension\Templating\Placeholder::class . ' instead.',
+    E_USER_DEPRECATED
+);
 
-/**
- * Helper for passing data between otherwise segregated Views. It's called
- * Placeholder to make its typical usage obvious, but can be used just as easily
- * for non-Placeholder things. That said, the support for this is only
- * guaranteed to effect subsequently rendered templates, and of course Layouts.
- *
- * @deprecated
- */
-class Placeholder extends AbstractHelper
-{
+class_exists(\Pimcore\Twig\Extension\Templating\Placeholder::class);
+
+if (false) {
     /**
-     * Registry key under which container registers itself
-     *
-     * @var string
+     * @deprecated since Pimcore 6.8, use Pimcore\Twig\Extension\Templating\Placeholder
      */
-    protected $_regKey = 'Placeholder';
+    class Placeholder extends \Pimcore\Twig\Extension\Templating\Placeholder {
 
-    /**
-     * @var Container[]
-     */
-    protected $containers = [];
-
-    public function getName()
-    {
-        return 'placeholder';
-    }
-
-    /**
-     * Retrieve object instance; optionally add meta tag
-     *
-     * @param string $containerName
-     *
-     * @return Container
-     */
-    public function __invoke($containerName = null)
-    {
-        $containerName = (string) $containerName;
-        if (empty($this->containers[$containerName])) {
-            $this->containers[$containerName] = $this->containerService->getContainer($this->_regKey . '_' . $containerName);
-        }
-
-        return $this->containers[$containerName];
     }
 }
