+++ lib/Templating/Helper/Placeholder/CacheBusterAware.php
@@ -1,4 +1,7 @@
 <?php
+
+declare(strict_types=1);
+
 /**
  * Pimcore
  *
@@ -14,38 +17,19 @@
 
 namespace Pimcore\Templating\Helper\Placeholder;
 
-/**
- * Class CacheBusterAware
- *
- * adds cache buster functionality to placeholder helper
- *
- * @deprecated
- */
-abstract class CacheBusterAware extends AbstractHelper
-{
-    /**
-     * @var bool
-     */
-    protected $cacheBuster = true;
+@trigger_error(
+    'Pimcore\Templating\Helper\Placeholder\CacheBusterAware is deprecated since version 6.8.0 and will be removed in 7.0.0. ' .
+    ' Use ' . \Pimcore\Twig\Extension\Templating\Placeholder\CacheBusterAware::class . ' instead.',
+    E_USER_DEPRECATED
+);
 
-    /**
-     * prepares entries with cache buster prefix
-     */
-    abstract protected function prepareEntries();
+class_exists(\Pimcore\Twig\Extension\Templating\Placeholder\CacheBusterAware::class);
 
+if (false) {
     /**
-     * @return bool
+     * @deprecated since Pimcore 6.8, use Pimcore\Twig\Extension\Templating\Placeholder\CacheBusterAware
      */
-    public function isCacheBuster()
-    {
-        return $this->cacheBuster;
-    }
+    abstract class CacheBusterAware extends \Pimcore\Twig\Extension\Templating\Placeholder\CacheBusterAware {
 
-    /**
-     * @param bool $cacheBuster
-     */
-    public function setCacheBuster($cacheBuster)
-    {
-        $this->cacheBuster = $cacheBuster;
     }
 }
