+++ lib/Templating/Helper/Traits/WebLinksTrait.php
@@ -17,80 +17,19 @@ declare(strict_types=1);
 
 namespace Pimcore\Templating\Helper\Traits;
 
-use Pimcore\Templating\Helper\WebLink;
+@trigger_error(
+    'Pimcore\Templating\Helper\Traits\WebLinksTrait is deprecated since version 6.8.0 and will be removed in 7.0.0. ' .
+    ' Use ' . \Pimcore\Twig\Extension\Templating\Traits\WebLinksTrait::class . ' instead.',
+    E_USER_DEPRECATED
+);
 
-trait WebLinksTrait
-{
+class_exists(\Pimcore\Twig\Extension\Templating\Traits\WebLinksTrait::class);
+
+if (false) {
     /**
-     * @var WebLink
+     * @deprecated since Pimcore 6.8, use Pimcore\Twig\Extension\Templating\Traits\WebLinksTrait
      */
-    protected $webLinkHelper;
-
-    /**
-     * Whether to use WebLinks (HTTP/2 push) for every item. Can be
-     * overridden on an item level.
-     *
-     * @var bool
-     */
-    protected $webLinksEnabled = false;
-
-    public function webLinksEnabled(bool $enabled = null)
-    {
-        if (null !== $enabled) {
-            $this->webLinksEnabled = $enabled;
-        }
-
-        return $this->webLinksEnabled;
-    }
-
-    public function enableWebLinks(): self
-    {
-        $this->webLinksEnabled(true);
-
-        return $this;
-    }
-
-    public function getWebLinkAttributes(): array
-    {
-        return $this->webLinkAttributes;
-    }
-
-    public function setWebLinkAttributes(array $webLinkAttributes)
-    {
-        $this->webLinkAttributes = $webLinkAttributes;
-    }
-
-    protected function handleWebLink(\stdClass $item, string $source, array $itemAttributes)
-    {
-        if (empty($source)) {
-            return;
-        }
-
-        if (!$this->webLinksEnabled && !isset($itemAttributes['webLink'])) {
-            return;
-        }
-
-        $attributes = $this->webLinkAttributes;
-        if (isset($itemAttributes['webLink'])) {
-            if (is_bool($itemAttributes['webLink'])) {
-                // set webLink to false to disable webLink on the item level. this allows to
-                // enable web links for the whole helper while disabling them for individual items
-                if (!$itemAttributes['webLink']) {
-                    return;
-                } else {
-                    $itemAttributes['webLink'] = [];
-                }
-            }
-
-            $attributes = array_merge($attributes, $itemAttributes['webLink']);
-        }
-
-        $method = 'preload';
-        if (isset($attributes['method'])) {
-            $method = $attributes['method'];
-            unset($attributes['method']);
-        }
-
-        call_user_func([$this->webLinkHelper, $method], $source, $attributes);
+    trait WebLinksTrait {
+        use \Pimcore\Twig\Extension\Templating\Traits\WebLinksTrait;
     }
 }
