+++ lib/Twig/Extension/Templating/Navigation/Exception/RendererNotFoundException.php
@@ -1,0 +1,28 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Pimcore
+ *
+ * This source file is available under two different licenses:
+ * - GNU General Public License version 3 (GPLv3)
+ * - Pimcore Enterprise License (PEL)
+ * Full copyright and license information is available in
+ * LICENSE.md which is distributed with this source code.
+ *
+ * @copyright  Copyright (c) Pimcore GmbH (http://www.pimcore.org)
+ * @license    http://www.pimcore.org/license     GPLv3 and PEL
+ */
+
+namespace Pimcore\Twig\Extension\Templating\Navigation\Exception;
+
+class RendererNotFoundException extends \InvalidArgumentException
+{
+    public static function create(string $name): self
+    {
+        return new static(sprintf('The navigation renderer "%s" was not found', $name));
+    }
+}
+
+class_alias(RendererNotFoundException::class, 'Pimcore\Templating\Helper\Navigation\Exception\RendererNotFoundException');
